requirements.txt
flet>=0.10.0
mysql-connector-python>=8.1.0